requirements.txt
Django>=5.1,<5.2
django-ninja>=1.3,<2.0
celery>=5.4,<6.0
redis>=5.0,<6.0
psycopg2-binary>=2.9,<3.0
Pillow>=10.0,<11.0
gunicorn>=22.0,<23.0
django-environ>=0.11,<1.0
cairosvg>=2.7,<3.0
django-allauth>=0.63,<1.0
requests>=2.31,<3.0
cryptography>=42.0,<43.0
